Fix bugs and make rest of the project clean

+++ libparsers/tests/simple_tests/test_math.cpp
@@ -8,7 +8,7 @@ void print_evaluate_curve(const openbve2::math::evaulate_curve_t& val) {
 }
 
 void test_math() {
-	auto const distance = float(12.56637f * (5.0f / 4.0f) * (1.0f / 0.7071067811f));
+	auto const distance = 12.56637f * (5.0f / 4.0f) * (1.0f / 0.7071067811f);
 	auto const vala =
 	    openbve2::math::evaluate_curve(glm::vec3(0), glm::vec3(0, -1, 1), distance, -10);
 	auto valb = openbve2::math::evaluate_curve(vala.position, vala.tangent, distance, -10);
